Remove "regular" option from Scorer._compute_acc
The "regular" option is not used and requires checking an unused field in the config. This option was previously removed for `Scorer._compute_f1` as a bug fix, so I am also removing it from `Scorer._compute_acc` to completely remove the field (otherwise the code still errors out). With this fix, the ReCoRD evaluation runs with error for me.